refactor: reorganized mocks
