--- tests/mockedModels.ts
@@ -1,93 +1,0 @@
-export const mockedPaybutton = {
-  id: 4,
-  providerUserId: 'mocked-uid',
-  name: 'mocked-name',
-  buttonData: 'mockedData',
-  uuid: '730bfa24-eb57-11ec-b722-0242ac150002',
-  createdAt: new Date('2022-05-27T15:18:42.000Z'),
-  updatedAt: new Date('2022-05-27T15:18:42.000Z'),
-  addresses: [
-    {
-      id: 1,
-      address: 'mockedaddress0nkus8hzv367za28j900c7tv5v8pc',
-      createdAt: new Date('2022-05-27T15:18:42.000Z'),
-      updatedAt: new Date('2022-05-27T15:18:42.000Z'),
-      chainId: 1,
-      paybuttonId: 1
-    },
-    {
-      id: 2,
-      address: 'mockedaddress0nkush83z76az28900c7tj5vpc8f',
-      createdAt: new Date('2022-05-27T15:18:42.000Z'),
-      updatedAt: new Date('2022-05-27T15:18:42.000Z'),
-      chainId: 2,
-      paybuttonId: 1
-    }
-  ]
-}
-
-export const mockedPaybuttonList = [
-  {
-    id: 1,
-    providerUserId: 'mocked-uid',
-    name: 'mocked-name-1',
-    buttonData: 'mockedData',
-    uuid: '730bfa24-eb57-11ec-b722-0242ac150002',
-    createdAt: new Date('2022-05-27T15:18:42.000Z'),
-    updatedAt: new Date('2022-05-27T15:18:42.000Z'),
-    addresses: [
-      {
-        id: 1,
-        address: 'mockedaddress0nkus8hzv367za28j900c7tv5v8pc',
-        createdAt: new Date('2022-05-27T15:18:42.000Z'),
-        updatedAt: new Date('2022-05-27T15:18:42.000Z'),
-        chainId: 1,
-        paybuttonId: 1
-      },
-      {
-        id: 2,
-        address: 'mockedaddress0nkush83z76az28900c7tj5vpc8f',
-        createdAt: new Date('2022-05-27T15:18:42.000Z'),
-        updatedAt: new Date('2022-05-27T15:18:42.000Z'),
-        chainId: 2,
-        paybuttonId: 1
-      }
-    ]
-  },
-  {
-    id: 2,
-    providerUserId: 'mocked-uid',
-    name: 'mocked-name-2',
-    buttonData: 'mockedData',
-    uuid: '133fb8aa-eb57-11ec-b722-0242ac150002',
-    createdAt: new Date('2022-05-27T15:18:42.000Z'),
-    updatedAt: new Date('2022-05-27T15:18:42.000Z'),
-    addresses: [
-      {
-        id: 3,
-        address: 'mockedaddress0nkus8hzv367za28j900c7tv5v8pc',
-        createdAt: new Date('2022-05-27T15:18:42.000Z'),
-        updatedAt: new Date('2022-05-27T15:18:42.000Z'),
-        chainId: 1,
-        paybuttonId: 2
-      },
-      {
-        id: 4,
-        address: 'mockedaddress0nkush83z76az28900c7tj5vpc8f',
-        createdAt: new Date('2022-05-27T15:18:42.000Z'),
-        updatedAt: new Date('2022-05-27T15:18:42.000Z'),
-        chainId: 2,
-        paybuttonId: 2
-      }
-    ]
-  }
-]
-
-export const mockedChain = {
-  id: 1,
-  slug: 'bitcoincash',
-  ticker: 'bch',
-  title: 'Bitcoin Cash',
-  createdAt: new Date('2022-05-27T15:18:42.000Z'),
-  updatedAt: new Date('2022-05-27T15:18:42.000Z')
-}
