+++ tests/unittests/bchdService.test.ts
@@ -1,108 +1,8 @@
-import {
-  Transaction,
-  UnspentOutput,
-  GetTransactionResponse,
-  GetAddressTransactionsResponse,
-  GetAddressUnspentOutputsResponse
-} from 'grpc-bchrpc-node'
 import rewire from 'rewire'
+import { mockedGrpc } from '../mockedObjects'
 const bchdService = rewire('../../services/bchdService')
 
-const unspentOutputFromObject = (obj: UnspentOutput.AsObject): UnspentOutput => {
-  const uo = new UnspentOutput()
-  uo.setPubkeyScript(obj.pubkeyScript)
-  uo.setValue(obj.value)
-  uo.setIsCoinbase(obj.isCoinbase)
-  uo.setBlockHeight(obj.blockHeight)
-  return uo
-}
-
-const transactionFromObject = (obj: Transaction.AsObject): Transaction => {
-  const t = new Transaction()
-  t.setHash(obj.hash)
-  t.setVersion(obj.version)
-  t.setLockTime(obj.lockTime)
-  t.setSize(obj.size)
-  t.setTimestamp(obj.timestamp)
-  t.setConfirmations(obj.confirmations)
-  t.setBlockHeight(obj.blockHeight)
-  t.setBlockHash(obj.blockHash)
-  return t
-}
-
-const grpcMock = {
-  getAddressTransactions: (_: object) => {
-    const res = new GetAddressTransactionsResponse()
-    const t1 = transactionFromObject({
-      hash: 'LUZSpMOab+ZYlyQNxF0XasKpArgQAX633LoA5CBPGgE=',
-      version: 1,
-      lockTime: 0,
-      size: 219,
-      timestamp: 1653460454,
-      confirmations: 60,
-      blockHeight: 741620,
-      blockHash: 'jzSPV4kkI3x5Fdoow/ei3f7Zit+oGMYCAAAAAAAAAAA=',
-      inputsList: [],
-      outputsList: []
-    })
-    const t2 = transactionFromObject({
-      hash: 'jiZHfE+AohEJglMO29nQ5aTR6F/n4Om2whzEZUiXcHk=',
-      version: 2,
-      lockTime: 0,
-      size: 225,
-      timestamp: 1653459437,
-      confirmations: 61,
-      blockHeight: 741619,
-      blockHash: 'A6kjJsl4gaVrY0Z15k0SoRzfKv0Fis8EAAAAAAAAAAA=',
-      inputsList: [],
-      outputsList: []
-    })
-    res.setConfirmedTransactionsList([t1, t2])
-    return res
-  },
-  getAddressUtxos: (_: object) => {
-    const res = new GetAddressUnspentOutputsResponse()
-    res.setOutputsList([
-      unspentOutputFromObject({
-        pubkeyScript: 'dqkUYF1GSq6KqSQSKPbQtcOJWRBPEFaIrA==',
-        value: 547,
-        isCoinbase: false,
-        blockHeight: 684161
-      }),
-      unspentOutputFromObject({
-        pubkeyScript: 'dqkUYF1GSq6KqSQSKPbQtcOJWRBPEFaIrA==',
-        value: 122,
-        isCoinbase: false,
-        blockHeight: 657711
-      }),
-      unspentOutputFromObject({
-        pubkeyScript: 'dqkUYF1GSq6KqSQSKPbQtcOJWRBPEFaIrA==',
-        value: 1111,
-        isCoinbase: false,
-        blockHeight: 596627
-      })
-    ])
-    return res
-  },
-  getTransaction: (_: object) => {
-    const res = new GetTransactionResponse()
-    res.setTransaction(transactionFromObject({
-      hash: 'hu9m3BZg/zlxis7ehc0x/+9qELXC8dkbimOtc5v598s=',
-      version: 2,
-      lockTime: 0,
-      size: 518,
-      timestamp: 1653653100,
-      confirmations: 0,
-      blockHeight: 0,
-      blockHash: '',
-      inputsList: [],
-      outputsList: []
-    }))
-    return res
-  }
-}
-
-bchdService.__set__('grpc', grpcMock)
+bchdService.__set__('grpc', mockedGrpc)
 
 describe('Test service returned objects consistency', () => {
   it('test getAddress', async () => {
@@ -113,7 +13,21 @@ describe('Test service returned objects consistency', () => {
           hash: 'LUZSpMOab+ZYlyQNxF0XasKpArgQAX633LoA5CBPGgE=',
           version: 1,
           inputsList: [],
-          outputsList: [],
+          outputsList: [{
+            index: 0,
+            value: 431247724,
+            pubkeyScript: 'dqkUeCxnbWKveaKpCSRhJC/poE+saL+IrA==',
+            address: 'qpuzcemdv2hhng4fpyjxzfp0axsyltrghutla9rfnm',
+            scriptClass: 'pubkeyhash',
+            disassembledScript: 'OP_DUP OP_HASH160 782c676d62af79a2a9092461242fe9a04fac68bf OP_EQUALVERIFY OP_CHECKSIG'
+          }, {
+            index: 1,
+            value: 227413293,
+            pubkeyScript: 'dqkUokKnAjaab8AVlPxzrrxk1aRq4BOIrA==',
+            address: 'qz3y9fczx6dxlsq4jn788t4uvn26g6hqzvrczjuzz2',
+            scriptClass: 'pubkeyhash',
+            disassembledScript: 'OP_DUP OP_HASH160 a242a702369a6fc01594fc73aebc64d5a46ae013 OP_EQUALVERIFY OP_CHECKSIG'
+          }],
           lockTime: 0,
           size: 219,
           timestamp: 1653460454,
